+++ game/CMakeLists.txt
@@ -7,4 +7,5 @@ file(GLOB game_src
 add_library(game SHARED ${game_src})
 target_include_directories(game PUBLIC include)
 target_include_directories(game PUBLIC include/physics)
+target_include_directories(game PUBLIC include/ai)
 target_link_libraries(game deltacore)
